+++ khorosjx/spaces.py
@@ -6,7 +6,7 @@
 :Example:        ``space_info = khorosjx.spaces.get_space_info(1234)``
 :Created By:     Jeff Shurtliff
 :Last Modified:  Jeff Shurtliff
-:Modified Date:  05 Dec 2019
+:Modified Date:  16 Dec 2019
 """
 
 from . import core, errors
@@ -119,3 +119,55 @@ def get_browse_id(space_id, return_type='str'):
     """
     browse_id = get_place_id(space_id, return_type)
     return browse_id
+
+
+# Define function to get the permitted content types for a space
+def get_permitted_content_types(id_value, id_type='browse_id', return_type='list'):
+    """This function returns the permitted content types for a given space.
+
+    :param id_value: The space identifier as a Browse ID (default), Place ID or Space ID
+    :type id_value: int, str
+    :param id_type: Determines if the ``id_value`` is a ``browse_id`` (Default), ``place_id`` or ``space_id``
+    :type id_type: str
+    :param return_type: Determines if the result should be returned in ``list`` (Default), ``tuple`` or ``str`` format
+    :type return_type: str
+    :returns: The permitted content types in list, tuple or string format
+    :raises: SpaceNotFountError, GETRequestError
+    """
+    # Verify that the core connection has been established
+    verify_core_connection()
+
+    # Get the appropriate ID for the space to check
+    accepted_id_types = ['browse_id', 'place_id', 'space_id']
+    if id_type in accepted_id_types:
+        if id_type != "browse_id" and id_type != "place_id":
+            id_value = get_browse_id(id_value)
+    else:
+        exception_msg = "The supplied lookup type for the API query is not recognized. (Examples of valid " + \
+                        "lookup types include 'browse_id' and 'space_id')"
+        raise errors.exceptions.InvalidLookupTypeError(exception_msg)
+
+    # Get the permitted content types
+    query_url = f"{base_url}/places/{id_value}/permissions"
+    space_permissions = core.get_request_with_retries(query_url, return_json=True)
+
+    # Check for an error in the response
+    try:
+        error_status_code = space_permissions['error']['status']
+        # Check if the error was because the space could not be found
+        if error_status_code == 404:
+            raise errors.exceptions.SpaceNotFoundError
+        else:
+            exception_msg = f"The API request failed with a {error_status_code} status code and the " + \
+                            f"following error: {space_permissions['error']['message']}"
+            raise errors.exceptions.GETRequestError(exception_msg)
+    except KeyError:
+        pass
+
+    # Get and return the permitted content types as a list or string
+    content_types = space_permissions['contentTypes']
+    if return_type == 'tuple':
+        content_types = tuple(content_types)
+    elif return_type == 'str':
+        content_types = ', '.join(content_types)
+    return content_types